+++ src/simulator/GreedySimulator.cpp
@@ -1296,7 +1296,7 @@ cda_rail::simulator::GreedySimulator::time_to_exit_objective(
 
 std::pair<double, double> cda_rail::simulator::GreedySimulator::get_ma_and_maxv(
     size_t tr, const std::vector<double>& train_velocities,
-    std::optional<double> next_stop, int h, int dt,
+    std::optional<size_t> next_stop, int h, int dt,
     const std::vector<std::pair<double, double>>&  train_positions,
     const std::unordered_set<size_t>&              trains_in_network,
     const std::unordered_set<size_t>&              trains_left,
